requirements.txt
flask 
# openai==0.28
together
python-dotenv

